move step legend to own file
